# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/marketplace": {
      "get": {
        "operationId": "MarketplaceController_listPublished",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Keyword matched against title/description. Max 200 characters.",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort direction. Default latest.",
            "schema": {
              "type": "string",
              "enum": [
                "latest",
                "oldest"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size, default 20, max 100.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MarketplaceItemDto"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List published marketplace items",
        "tags": [
          "marketplace"
        ]
      },
      "post": {
        "operationId": "MarketplaceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMarketplaceItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceItemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a marketplace item (admin)",
        "tags": [
          "marketplace"
        ]
      }
    },
    "/api/marketplace/{slug}": {
      "get": {
        "operationId": "MarketplaceController_getPublishedBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceItemDetailDto"
                }
              }
            }
          }
        },
        "summary": "Read a published marketplace item by slug",
        "tags": [
          "marketplace"
        ]
      }
    },
    "/api/marketplace/{id}": {
      "patch": {
        "operationId": "MarketplaceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMarketplaceItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceItemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a marketplace item (admin)",
        "tags": [
          "marketplace"
        ]
      }
    },
    "/api/marketplace/{id}/publish": {
      "post": {
        "operationId": "MarketplaceController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceItemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Publish a marketplace item (admin)",
        "tags": [
          "marketplace"
        ]
      }
    },
    "/api/marketplace/{id}/unpublish": {
      "post": {
        "operationId": "MarketplaceController_unpublish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceItemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Unpublish a marketplace item (admin)",
        "tags": [
          "marketplace"
        ]
      }
    },
    "/api/marketplace/{id}/media": {
      "post": {
        "operationId": "MarketplaceController_attachMedia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachMarketplaceMediaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceMediaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Attach media to a marketplace item (admin)",
        "tags": [
          "marketplace"
        ]
      }
    },
    "/api/marketplace/{id}/media/{mediaId}": {
      "delete": {
        "operationId": "MarketplaceController_detachMedia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Detach media from a marketplace item (admin)",
        "tags": [
          "marketplace"
        ]
      }
    },
...
    "/api/admin/media/upload-url": {
      "post": {
        "operationId": "MediaController_createUploadUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string",
                      "description": "Server-generated object key."
                    },
                    "uploadUrl": {
                      "type": "string",
                      "description": "Presigned PUT URL."
                    },
                    "expiresIn": {
                      "type": "integer",
                      "description": "URL validity in seconds."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a presigned upload URL (admin)",
        "tags": [
          "media"
        ]
      }
    },
    "/api/admin/media/confirm": {
      "post": {
        "operationId": "MediaController_confirm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string"
                    },
                    "confirmed": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Confirm a completed upload (admin)",
        "tags": [
          "media"
        ]
      }
    },
    "/api/admin/media/read-url": {
      "post": {
        "operationId": "MediaController_readUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "readUrl": {
                      "type": "string",
                      "description": "Presigned GET URL."
                    },
                    "expiresIn": {
                      "type": "integer",
                      "description": "URL validity in seconds."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a presigned read URL for private media (admin)",
        "tags": [
          "media"
        ]
      }
    },
...
  },
  "info": {
    "title": "Xolvon Backend API",
    "description": "HTTP contract for the Xolvon backend (global prefix /api). Auth: `Authorization: Bearer <accessToken>` (JWT, issued by POST /api/auth/register and POST /api/auth/login). Public routes need no credentials; routes listed with bearer security reject missing or invalid tokens with 401.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token. Send it as `Authorization: Bearer <accessToken>`."
      }
    },
    "schemas": {
...
      "MarketplaceMediaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Media id (UUID v4)."
          },
          "objectKey": {
            "type": "string"
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "image",
              "video",
              "deck"
            ]
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "objectKey",
          "mediaType",
          "sortOrder"
        ]
      },
      "MarketplaceItemDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Item id (UUID v4)."
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "externalUrl": {
            "type": "string",
            "description": "Outbound https product URL."
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarketplaceMediaDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "description",
          "capabilities",
          "externalUrl",
          "status",
          "media"
        ]
      },
      "CreateMarketplaceItemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Item title, max 200 characters.",
            "maxLength": 200
          },
          "slug": {
            "type": "string",
            "description": "URL slug: lowercase letters, digits, hyphens. Max 120 characters.",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "description": "Description, max 2000 characters, sanitized on write.",
            "maxLength": 2000
          },
          "externalUrl": {
            "type": "string",
            "description": "Outbound product URL, https-only, max 2048 characters.",
            "maxLength": 2048
          },
          "capabilities": {
            "description": "Capability tags, max 20 items.",
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "slug",
          "externalUrl"
        ]
      },
      "MarketplaceItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Item id (UUID v4)."
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "externalUrl": {
            "type": "string",
            "description": "Outbound https product URL."
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "description",
          "capabilities",
          "externalUrl",
          "status"
        ]
      },
      "UpdateMarketplaceItemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Item title, max 200 characters.",
            "maxLength": 200
          },
          "slug": {
            "type": "string",
            "description": "URL slug: lowercase letters, digits, hyphens. Max 120 characters.",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "description": "Description, max 2000 characters, sanitized on write.",
            "maxLength": 2000
          },
          "externalUrl": {
            "type": "string",
            "description": "Outbound product URL, https-only, max 2048 characters.",
            "maxLength": 2048
          },
          "capabilities": {
            "description": "Capability tags, max 20 items.",
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AttachMarketplaceMediaDto": {
        "type": "object",
        "properties": {
          "objectKey": {
            "type": "string",
            "description": "R2 object key from the upload flow, max 1024 characters.",
            "maxLength": 1024
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "image",
              "video",
              "deck"
            ]
          },
          "sortOrder": {
            "type": "number",
            "description": "Display order, min 0.",
            "minimum": 0
          }
        },
        "required": [
          "objectKey",
          "mediaType"
        ]
      },
...
      "CreateUploadUrlDto": {
        "type": "object",
        "properties": {
          "prefix": {
            "type": "string",
            "enum": [
              "public/site",
              "public/projects",
              "private/projects",
              "private/courses"
            ],
            "description": "Allowed storage prefix for the key."
          },
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp",
              "application/pdf",
              "video/mp4"
            ],
            "description": "MIME type of the file being uploaded."
          },
          "size": {
            "type": "number",
            "description": "File size in bytes, 1-524288000.",
            "minimum": 1,
            "maximum": 524288000
          },
          "filename": {
            "type": "string",
            "description": "Original file name."
          }
        },
        "required": [
          "prefix",
          "contentType",
          "size",
          "filename"
        ]
      },
      "ConfirmMediaDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Object key of the uploaded file."
          }
        },
        "required": [
          "key"
        ]
      },
...
    }
  }
}